Added custom messages and created a python file with all the messages.
Custom.xml contains all of our custom messages, we just need to alter
that then run generator.py to create a new one and overwrite
examples.py. Added in the udp host and port location in bwtest so it
knows where the BBB is located, although we still need to transmit data
from it properly.